修正数据导出逻辑，处理无法写入问题 添加数据导入失败时toast 修正书架页面书本顺序颠倒 修复其他界面切换到阅读中时会在空页和内容之间闪烁 添加点开章节选择器时默认展开当前章节所在卷，并滚动到该元素 修复【更新可用】【添加至书架】等对话框关闭时不应出现的延迟 点击翻页的判定线向左略微调整 书架页点击书本后退出时还原滚动进度和选择项 将导出的数据文件从文本格式变为zip格式 取消分享时用户需手动选择文件缓存目录，自动将文件生成至缓存目录 添加批量添加书本至书架功能 添加书架分享功能

+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/data/work/ExportDataWork.kt
@@ -23,6 +23,8 @@ import java.io.File
 import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import java.io.IOException
+import java.util.zip.ZipEntry
+import java.util.zip.ZipOutputStream
 
 @HiltWorker
 class ExportDataWork @AssistedInject constructor(
@@ -74,18 +76,21 @@ class ExportDataWork @AssistedInject constructor(
             }
             .build()
             .toJson()
-        return creatFile(fileUri, json)
+        return createFile(fileUri, json)
     }
 
-    private fun creatFile(fileUri: Uri, json: String): Result {
+    private fun createFile(fileUri: Uri, json: String): Result {
         return try {
             File(applicationContext.filesDir, "data").let { if (!it.exists()) it.mkdir() }
-            val file = fileUri.toFile()
-            if (file.exists())
-                file.delete()
-            file.createNewFile()
+            if (fileUri.scheme.equals("file")) {
+                val file = fileUri.toFile()
+                if (file.exists())
+                    file.delete()
+                file.createNewFile()
+            }
             applicationContext.contentResolver.openFileDescriptor(fileUri, "w")?.use { parcelFileDescriptor ->
-                FileOutputStream(parcelFileDescriptor.fileDescriptor).use {
+                ZipOutputStream(FileOutputStream(parcelFileDescriptor.fileDescriptor)).use {
+                    it.putNextEntry(ZipEntry("data.json"))
                     it.write(json.toByteArray())
                 }
             }
